refactor: Remove useless paths code from import statement
